+++ Pawn Editor Forked/Source/PawnEditorForked/UI/PawnList.cs
@@ -62,7 +62,9 @@ public static partial class PawnEditor
                     selectedCategory == PawnCategory.Humans ? default : new Vector3(-0.01f, 0, 0),
                     1 / Mathf.Clamp(pawn.BodySize, 1, 5));
             if (portraitTex != null)
+            {
                 GUI.DrawTexture(new(105f - portraitSize.x / 2f, 40f - portraitSize.y / 2f, portraitSize.x, portraitSize.y), portraitTex);
+            }
             GUI.color = Color.white;
             var label = pawn.Name is NameTriple nameTriple ? nameTriple.Nick.NullOrEmpty() ? nameTriple.First : nameTriple.Nick : pawn.LabelShort;
             using (new TextBlock(TextAnchor.MiddleLeft))